# src/main/java/com/parkar/parksaathi/dto/request/RefreshTokenRequest.java
package com.parkar.parksaathi.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class RefreshTokenRequest {

    @NotBlank(message = "Refresh token is required")
    private String refreshToken;
}